move into seperate commands in prep for dump command
