+++ package.json
@@ -31,7 +31,7 @@
     "ajv": "^6.5.2",
     "auth0": "^2.11.0",
     "auth0-extension-tools": "^1.2.1",
-    "auth0-source-control-extension-tools": "stevezau/auth0-source-control-extension-tools#improve_automation",
+    "auth0-source-control-extension-tools": "^3.0.0",
     "http-proxy-agent": "^2.0.0",
     "https-proxy-agent": "^2.0.0",
     "js-yaml": "^3.12.0",
