+++ src/commands/index.js
@@ -1,0 +1,7 @@
+import deploy from './deploy';
+import dump from './dump';
+
+export default {
+  deploy,
+  dump
+};
